protect api routes

+++ app/api/essay/[id]/route.ts
@@ -1,25 +1,29 @@
+import { auth } from '@/auth'
 import { db } from '@/database'
 import { essay } from '@/database/schema'
 import { eq, sql } from 'drizzle-orm'
 import { NextRequest, NextResponse } from 'next/server'
 
-export async function DELETE(req: NextRequest) {
+export const DELETE = auth(async function DELETE(req) {
+    if (!req.auth) {
+        return NextResponse.json(
+            {
+                message: 'User unauthenticated',
+            },
+            { status: 401 }
+        )
+    }
     const url = new URL(req.url)
     const essayID = Number(url.pathname.split('/').pop())
 
     if (isNaN(essayID)) {
-        return NextResponse.json(
-            { success: false, message: 'Invalid ID' },
-            { status: 400 }
-        )
+        return NextResponse.json({ success: false, message: 'Invalid ID' }, { status: 400 })
     }
 
     try {
         await db.transaction(async (tx) => {
             // DELETE THE CHILD DATA TABLE FIRST TO AVOID FK CONSTRAINTS ERROR
-            await tx.execute(
-                sql`DELETE FROM essay_evaluations WHERE essay_id = ${essayID}`
-            )
+            await tx.execute(sql`DELETE FROM essay_evaluations WHERE essay_id = ${essayID}`)
             await tx.execute(sql`DELETE FROM essay_summaries WHERE essay_id = ${essayID}`)
 
             await tx.delete(essay).where(eq(essay.id, essayID))
@@ -28,9 +32,6 @@ export async function DELETE(req: NextRequest) {
         return NextResponse.json({ success: true }, { status: 200 })
     } catch (err) {
         console.error('Failed to delete rubric:', err)
-        return NextResponse.json(
-            { success: false, message: 'Server error' },
-            { status: 500 }
-        )
+        return NextResponse.json({ success: false, message: 'Server error' }, { status: 500 })
     }
-}
+})
